new request

+++ app/Http/Controllers/UserController.php
@@ -1,83 +1,48 @@
-<?php namespace App\Http\Controllers;
+<?php
 
+namespace App\Http\Controllers;
+
+use App\Models\User;
+use App\Ninja\Mailers\ContactMailer;
+use App\Ninja\Mailers\UserMailer;
+use App\Ninja\Repositories\AccountRepository;
+use App\Services\UserService;
 use Auth;
-use Config;
-use Datatable;
-use DB;
-use Event;
 use Input;
-use View;
-use Request;
+use Password;
 use Redirect;
+use Request;
 use Session;
 use URL;
 use Utils;
 use Validator;
-use Illuminate\Auth\Passwords\TokenRepositoryInterface;
-use App\Models\User;
-use App\Http\Requests;
-use App\Ninja\Repositories\AccountRepository;
-use App\Ninja\Mailers\ContactMailer;
-use App\Ninja\Mailers\UserMailer;
+use View;
 
 class UserController extends BaseController
 {
     protected $accountRepo;
     protected $contactMailer;
     protected $userMailer;
+    protected $userService;
 
-    public function __construct(AccountRepository $accountRepo, ContactMailer $contactMailer, UserMailer $userMailer)
+    public function __construct(AccountRepository $accountRepo, ContactMailer $contactMailer, UserMailer $userMailer, UserService $userService)
     {
-        parent::__construct();
+        //parent::__construct();
 
         $this->accountRepo = $accountRepo;
         $this->contactMailer = $contactMailer;
         $this->userMailer = $userMailer;
+        $this->userService = $userService;
+    }
+
+    public function index()
+    {
+        return Redirect::to('settings/' . ACCOUNT_USER_MANAGEMENT);
     }
 
     public function getDatatable()
     {
-        $query = DB::table('users')
-                  ->where('users.account_id', '=', Auth::user()->account_id);
-
-        if (!Session::get('show_trash:user')) {
-            $query->where('users.deleted_at', '=', null);
-        }
-
-        $query->where('users.public_id', '>', 0)
-              ->select('users.public_id', 'users.first_name', 'users.last_name', 'users.email', 'users.confirmed', 'users.public_id', 'users.deleted_at');
-
-        return Datatable::query($query)
-        ->addColumn('first_name', function ($model) { return link_to('users/'.$model->public_id.'/edit', $model->first_name.' '.$model->last_name); })
-        ->addColumn('email', function ($model) { return $model->email; })
-        ->addColumn('confirmed', function ($model) { return $model->deleted_at ? trans('texts.deleted') : ($model->confirmed ? trans('texts.active') : trans('texts.pending')); })
-        ->addColumn('dropdown', function ($model) {
-          $actions = '<div class="btn-group tr-action" style="visibility:hidden;">
-              <button type="button" class="btn btn-xs btn-default dropdown-toggle" data-toggle="dropdown">
-                '.trans('texts.select').' <span class="caret"></span>
-              </button>
-              <ul class="dropdown-menu" role="menu">';
-
-          if ($model->deleted_at) {
-              $actions .= '<li><a href="'.URL::to('restore_user/'.$model->public_id).'">'.uctrans('texts.restore_user').'</a></li>';
-          } else {
-              $actions .= '<li><a href="'.URL::to('users/'.$model->public_id).'/edit">'.uctrans('texts.edit_user').'</a></li>';
-
-              if (!$model->confirmed) {
-                  $actions .= '<li><a href="'.URL::to('send_confirmation/'.$model->public_id).'">'.uctrans('texts.send_invite').'</a></li>';
-              }
-
-              $actions .= '<li class="divider"></li>
-                <li><a href="javascript:deleteUser('.$model->public_id.')">'.uctrans('texts.delete_user').'</a></li>';
-          }
-
-           $actions .= '</ul>
-          </div>';
-
-          return $actions;
-        })
-        ->orderColumns(['first_name', 'email', 'confirmed'])
-        ->make();
+        return $this->userService->getDatatable(Auth::user()->account_id);
     }
 
     public function setTheme()
@@ -103,14 +68,14 @@ class UserController extends BaseController
     public function edit($publicId)
     {
         $user = User::where('account_id', '=', Auth::user()->account_id)
-                        ->where('public_id', '=', $publicId)->firstOrFail();
+                        ->where('public_id', '=', $publicId)
+                        ->withTrashed()
+                        ->firstOrFail();
 
         $data = [
-            'showBreadcrumbs' => false,
             'user' => $user,
             'method' => 'PUT',
             'url' => 'users/'.$publicId,
-            'title' => trans('texts.edit_user'),
         ];
 
         return View::make('users.edit', $data);
@@ -127,73 +92,71 @@ class UserController extends BaseController
     }
 
     /**
-     * Displays the form for account creation
-     *
+     * Displays the form for account creation.
      */
     public function create()
     {
-        if (!Auth::user()->registered) {
+        if (! Auth::user()->registered) {
             Session::flash('error', trans('texts.register_to_add_user'));
-            return Redirect::to('company/advanced_settings/user_management');
-        }        
-        if (!Auth::user()->confirmed) {
-            Session::flash('error', trans('texts.confirmation_required'));
-            return Redirect::to('company/advanced_settings/user_management');
+
+            return Redirect::to('settings/' . ACCOUNT_USER_MANAGEMENT);
         }
 
-        if (Utils::isNinja()) {
-            $count = User::where('account_id', '=', Auth::user()->account_id)->count();
-            if ($count >= MAX_NUM_USERS) {
-                Session::flash('error', trans('texts.limit_users'));
+        if (! Auth::user()->confirmed) {
+            Session::flash('error', trans('texts.confirmation_required'));
 
-                return Redirect::to('company/advanced_settings/user_management');
-            }
+            return Redirect::to('settings/' . ACCOUNT_USER_MANAGEMENT);
+        }
+
+        if (Utils::isNinja() && ! Auth::user()->caddAddUsers()) {
+            Session::flash('error', trans('texts.max_users_reached'));
+
+            return Redirect::to('settings/' . ACCOUNT_USER_MANAGEMENT);
         }
 
         $data = [
-          'showBreadcrumbs' => false,
           'user' => null,
           'method' => 'POST',
           'url' => 'users',
-          'title' => trans('texts.add_user'),
         ];
 
         return View::make('users.edit', $data);
     }
 
-    public function delete()
+    public function bulk()
     {
-        $userPublicId = Input::get('userPublicId');
+        $action = Input::get('bulk_action');
+        $id = Input::get('bulk_public_id');
+
         $user = User::where('account_id', '=', Auth::user()->account_id)
-                    ->where('public_id', '=', $userPublicId)->firstOrFail();
+                    ->where('public_id', '=', $id)
+                    ->withTrashed()
+                    ->firstOrFail();
 
-        $user->delete();
+        if ($action === 'archive') {
+            $user->delete();
+        } else {
+            if (! Auth::user()->caddAddUsers()) {
+                return Redirect::to('settings/' . ACCOUNT_USER_MANAGEMENT)
+                    ->with('error', trans('texts.max_users_reached'));
+            }
 
-        Session::flash('message', trans('texts.deleted_user'));
+            $user->restore();
+        }
 
-        return Redirect::to('company/advanced_settings/user_management');
-    }
+        Session::flash('message', trans("texts.{$action}d_user"));
 
-    public function restoreUser($userPublicId)
-    {
-        $user = User::where('account_id', '=', Auth::user()->account_id)
-                    ->where('public_id', '=', $userPublicId)
-                    ->withTrashed()->firstOrFail();
-
-        $user->restore();
-
-        Session::flash('message', trans('texts.restored_user'));
-
-        return Redirect::to('company/advanced_settings/user_management');
+        return Redirect::to('settings/' . ACCOUNT_USER_MANAGEMENT);
     }
 
     /**
-     * Stores new account
+     * Stores new account.
      *
+     * @param mixed $userPublicId
      */
     public function save($userPublicId = false)
     {
-        if (Auth::user()->account->isPro()) {
+        if (Auth::user()->hasFeature(FEATURE_USERS)) {
             $rules = [
                 'first_name' => 'required',
                 'last_name' => 'required',
@@ -201,7 +164,9 @@ class UserController extends BaseController
 
             if ($userPublicId) {
                 $user = User::where('account_id', '=', Auth::user()->account_id)
-                            ->where('public_id', '=', $userPublicId)->firstOrFail();
+                            ->where('public_id', '=', $userPublicId)
+                            ->withTrashed()
+                            ->firstOrFail();
 
                 $rules['email'] = 'required|email|unique:users,email,'.$user->id.',id';
             } else {
@@ -219,6 +184,10 @@ class UserController extends BaseController
                 $user->last_name = trim(Input::get('last_name'));
                 $user->username = trim(Input::get('email'));
                 $user->email = trim(Input::get('email'));
+                if (Auth::user()->hasFeature(FEATURE_USER_PERMISSIONS)) {
+                    $user->is_admin = boolval(Input::get('is_admin'));
+                    $user->permissions = Input::get('permissions');
+                }
             } else {
                 $lastUser = User::withTrashed()->where('account_id', '=', Auth::user()->account_id)
                             ->orderBy('public_id', 'DESC')->first();
@@ -230,14 +199,18 @@ class UserController extends BaseController
                 $user->username = trim(Input::get('email'));
                 $user->email = trim(Input::get('email'));
                 $user->registered = true;
-                $user->password = str_random(RANDOM_KEY_LENGTH);
-                $user->confirmation_code = str_random(RANDOM_KEY_LENGTH);
+                $user->password = strtolower(str_random(RANDOM_KEY_LENGTH));
+                $user->confirmation_code = strtolower(str_random(RANDOM_KEY_LENGTH));
                 $user->public_id = $lastUser->public_id + 1;
+                if (Auth::user()->hasFeature(FEATURE_USER_PERMISSIONS)) {
+                    $user->is_admin = boolval(Input::get('is_admin'));
+                    $user->permissions = Input::get('permissions');
+                }
             }
 
             $user->save();
 
-            if (!$user->confirmed) {
+            if (! $user->confirmed && Input::get('action') === 'email') {
                 $this->userMailer->sendConfirmation($user, Auth::user());
                 $message = trans('texts.sent_invite');
             } else {
@@ -246,8 +219,8 @@ class UserController extends BaseController
 
             Session::flash('message', $message);
         }
-        
-        return Redirect::to('company/advanced_settings/user_management');
+
+        return Redirect::to('users/' . $user->public_id . '/edit');
     }
 
     public function sendConfirmation($userPublicId)
@@ -258,40 +231,43 @@ class UserController extends BaseController
         $this->userMailer->sendConfirmation($user, Auth::user());
         Session::flash('message', trans('texts.sent_invite'));
 
-        return Redirect::to('company/advanced_settings/user_management');
+        return Redirect::to('settings/' . ACCOUNT_USER_MANAGEMENT);
     }
 
-
     /**
-     * Attempt to confirm account with code
+     * Attempt to confirm account with code.
      *
      * @param string $code
      */
-    public function confirm($code, TokenRepositoryInterface $tokenRepo)
+    public function confirm($code)
     {
         $user = User::where('confirmation_code', '=', $code)->get()->first();
-            
+
         if ($user) {
-            $notice_msg = trans('texts.security.confirmation');
+            $notice_msg = trans('texts.security_confirmation');
 
             $user->confirmed = true;
             $user->confirmation_code = '';
             $user->save();
 
             if ($user->public_id) {
-                //Auth::login($user);
-                $token = $tokenRepo->create($user);
+                Auth::logout();
+                $token = Password::getRepository()->create($user);
 
                 return Redirect::to("/password/reset/{$token}");
             } else {
-                if (Session::has(REQUESTED_PRO_PLAN)) {
-                    Session::forget(REQUESTED_PRO_PLAN);
-                    $invitation = $this->accountRepo->enableProPlan();
-
-                    return Redirect::to($invitation->getLink());
+                if (Auth::check()) {
+                    if (Session::has(REQUESTED_PRO_PLAN)) {
+                        Session::forget(REQUESTED_PRO_PLAN);
+                        $url = '/settings/account_management?upgrade=true';
+                    } else {
+                        $url = '/dashboard';
+                    }
                 } else {
-                    return Redirect::to(Auth::check() ? '/dashboard' : '/login')->with('message', $notice_msg);
+                    $url = '/login';
                 }
+
+                return Redirect::to($url)->with('message', $notice_msg);
             }
         } else {
             $error_msg = trans('texts.security.wrong_confirmation');
@@ -300,34 +276,12 @@ class UserController extends BaseController
         }
     }
 
-    /**
-     * Log the user out of the application.
-     *
-     */
-    /*
-    public function logout()
-    {
-        if (Auth::check()) {
-            if (!Auth::user()->registered) {
-                $account = Auth::user()->account;
-                $this->accountRepo->unlinkAccount($account);
-                $account->forceDelete();
-            }
-        }
-
-        Auth::logout();
-        Session::flush();
-
-        return Redirect::to('/')->with('clearGuestKey', true);
-    }
-    */
-    
     public function changePassword()
     {
         // check the current password is correct
-        if (!Auth::validate([
+        if (! Auth::validate([
             'email' => Auth::user()->email,
-            'password' => Input::get('current_password')
+            'password' => Input::get('current_password'),
         ])) {
             return trans('texts.password_error_incorrect');
         }
@@ -348,12 +302,12 @@ class UserController extends BaseController
         return RESULT_SUCCESS;
     }
 
-    public function switchAccount($newUserId) 
+    public function switchAccount($newUserId)
     {
         $oldUserId = Auth::user()->id;
         $referer = Request::header('referer');
         $account = $this->accountRepo->findUserAccounts($newUserId, $oldUserId);
-        
+
         if ($account) {
             if ($account->hasUserId($newUserId) && $account->hasUserId($oldUserId)) {
                 Auth::loginUsingId($newUserId);
@@ -364,8 +318,30 @@ class UserController extends BaseController
                 Session::put('_token', str_random(40));
             }
         }
-        
-        return Redirect::to($referer);
+
+        // If the user is looking at an entity redirect to the dashboard
+        preg_match('/\/[0-9*][\/edit]*$/', $referer, $matches);
+        if (count($matches)) {
+            return Redirect::to('/dashboard');
+        } else {
+            return Redirect::to($referer);
+        }
+    }
+
+    public function viewAccountByKey($accountKey)
+    {
+        $user = $this->accountRepo->findUser(Auth::user(), $accountKey);
+
+        if (! $user) {
+            return redirect()->to('/');
+        }
+
+        Auth::loginUsingId($user->id);
+        Auth::user()->account->loadLocalizationSettings();
+
+        $redirectTo = request()->redirect_to ?: '/';
+
+        return redirect()->to($redirectTo);
     }
 
     public function unlinkAccount($userAccountId, $userId)
@@ -377,6 +353,25 @@ class UserController extends BaseController
         Session::put(SESSION_USER_ACCOUNTS, $users);
 
         Session::flash('message', trans('texts.unlinked_account'));
-        return Redirect::to($referer);
+
+        return Redirect::to('/manage_companies');
+    }
+
+    public function manageCompanies()
+    {
+        return View::make('users.account_management');
+    }
+
+    public function saveSidebarState()
+    {
+        if (Input::has('show_left')) {
+            Session::put(SESSION_LEFT_SIDEBAR, boolval(Input::get('show_left')));
+        }
+
+        if (Input::has('show_right')) {
+            Session::put(SESSION_RIGHT_SIDEBAR, boolval(Input::get('show_right')));
+        }
+
+        return RESULT_SUCCESS;
     }
 }
